Cleaned up web3 context.
